用户id类型

+++ net/package.h
@@ -33,7 +33,7 @@ struct PackageHead
 {
     std::uint16_t head_size      = sizeof(PackageHead);
     std::uint16_t data_size      = 0;
-    std::uint32_t uid            = 0;
+    std::uint64_t uid            = 0;
     std::uint16_t cmd            = 0;
     std::uint16_t src            = 0; //from server ID
     std::uint16_t dst           = 0; //destination is server ID
@@ -59,14 +59,14 @@ public:
 
     void set_head_size(std::uint16_t head_size) {head_.head_size = head_size;}
     void set_data_size(std::uint16_t data_size) {head_.data_size = data_size;}
-    void set_uid(std::uint32_t uid) {head_.uid = uid;}
+    void set_uid(std::uint64_t uid) {head_.uid = uid;}
     void set_cmd(std::uint16_t cmd) {head_.cmd = cmd;}
     void set_src(std::uint16_t src) {head_.src = src;}
     void set_dst(std::uint16_t dst) {head_.dst = dst;}
 
     std::uint16_t head_size() const {return head_.head_size;}
     std::uint16_t data_size() const {return head_.data_size;}
-    std::uint32_t uid() const {return head_.uid;}
+    std::uint64_t uid() const {return head_.uid;}
     std::uint16_t cmd() const {return head_.cmd;}
     std::uint16_t src() const {return head_.src;}
     std::uint16_t dst() const {return head_.dst;}
